Module ruma_identity_service_api::association

Endpoints to create associations with a Matrix ID on the identity server.

Modules

bind_3pid

Endpoint to bind a session to a Matrix user ID.

check_3pid_validity

Endpoint to determine the validity of a 3PID.

email

Endpoints to create an association between a Matrix ID and an email ID on the identity server.

msisdn

Endpoints to create association between a phone number and a Matrix ID on the identity server.

unbind_3pid

Endpoint to remove an association between a session and a Matrix user ID.